+++ src/serve/tests/helpers.rs
@@ -30,6 +30,7 @@ pub(super) fn app_service() -> ServiceConfig {
         hook: Vec::new(),
         health_path: None,
         health_statuses: None,
+        restart: None,
         localhost_tls: false,
         octane: false,
         octane_workers: None,